language_model: Create provider module
