+++ crates/language_models/src/provider/anthropic/telemetry.rs
@@ -2,7 +2,7 @@ use anthropic::ANTHROPIC_API_URL;
 use anyhow::{Context as _, anyhow};
 use gpui::BackgroundExecutor;
 use http_client::{AsyncBody, HttpClient, Method, Request as HttpRequest};
-use language_model::{ANTHROPIC_PROVIDER_ID, LanguageModel};
+use language_model::{LanguageModel, provider::ANTHROPIC_PROVIDER_ID};
 use std::env;
 use std::sync::Arc;
 use util::ResultExt;
